[DM-15868] Work around long columns.
This gets me around some of the really long column names, which
Oracle doesn't want to deal with (Oracle has a 30 character limit
for column names).  If I change the name of the column, this is
in a few places and has to be consistent.

Also clean up a few places around these functions and group them.