[DM-15868] Add new images for each backend.
I could use the standard images, and have to mount PVs on them,
but that seems annoying for a few KBs.  Just bake it into the
images.  Update the build script and make a push script to
get the images to docker hub where k8s can suck them in.