[DM-16383] Add helm values for presto.
I'm using this chart: https://github.com/lotuc/presto-chart

These values configure the different backends correctly, which
for some reason the stable chart doesn't do.  Without an editor
on the images it can be really hard to add this, so I'm using
this chart which fixes this (but for some reason wasn't accepted
to charts/stable).