[DM-16383] Workaround for catalog.schema.table calls.
Because the JSQLParser CADC is using doesn't support the
catalog.schema.table calls (it's just too old of a version),
Patrick Dowler suggested that I use the table name rewriter
to rewrite a schema.table calls to catalog.schema.table using
the TableNameConverter.  This is done in the AdqlQueryImpl change.

Also, change the UWS tap_schema to go back to schema.table, and
add the wise metadata.  The wise metadata was generated with the
felis tool, along with some various fixes:

TABLES as the type needs to be lowercase, otherwise things are unhappy.
You can't load /tap/tables, or do any queries, and it tells you that
TABLE is an illegal data type.

Single quotes were not properly escaped so I escaped them in one of the
column descriptions where it uses quotes.

Also changed the names of the tables to use the schemas11, tables11,
columns11 tables.