[DM-16711] Play with tomcat connectionpool settings and commits.
Because there seem to be hanging transacations that are timing out,
I'm changing a few parameters.

1) Adjust pool to have more than 1 connection in it.  This will help
(and be required) because if there's a longer running query we might
run out of connections the way it was previously configured.

2) Presto complains about transactions timing out, and I noticed some
interesting setting of autocommit in QServQueryRunner, which I've
commented out and just forced to true in the context.xml.

With these settings I no longer see the error message in the
original ticket at all, but I'm not sure I haven't broken something
else.  I will be contacting CADC as to why they are playing with
autocommit like this.