[DM-17021] Disable table renaming.
If we aren't using presto, we don't need / shouldn't do this
remapping.  Adding the qserv to the name is to tell presto which
backend to route to.  If we want to go back to presto, we'll need
this until we get the three name call working natively (which
requires CADC upstream to take a better sql parser).