[DM-17455] Fix availability endpoint.
The availability endpoint was giving a SQL exception running
the query against TAP_SCHEMA.  This is because this goes through
a different call path that doesn't lowercase all the table names,
and mysql is sadly case sensitive as to table names.  So make
the table names lowercase, and get rid of some of the additional
ALMA checking.

Also get rid of some trailing whitespace in the context.xml file.

With this, the availability endpoint says everything is well.