[DM-17565] Fix ingress rules for redirects.
When making a sync TAP request, the server will respond with a
redirect to the result.  For the TAP service, sitting behind the
proxy, the URL is wrong (it doesn't have /api) and the scheme
may also be wrong (because it doesn't know it's sitting behind
something terminating SSL).

For the dev instance, we take the URL in the redirect (which
always starts with http, because it doesn't know if it may be
terminated), and rewrite it to add the /api in front of it.

For the int and stable instances, I've decided to force SSL on
(using SSL-redirect: true), because when rewriting the redirect
URL, the backend will always send an http URL, but it may need
to be https, if that was the endpoint that started the request.

Since we'll be turning on auth in the near future, it's probably
better to just force everyone over to HTTPS now.