[DM-17187] Get rid of monkeypatch, go back to mainline cadc-adql.
Bump up the dependency version of cadc-adql (which is part of the
https://github.com/opencadc/tap repo) to 1.1.4, which includes
my PR to add support for the DISTANCE function.  We need this
to have the DISTANCE parsing call back into the QServ specific
region converter.

Since we're going to 1.1.4, which has some of pdowler's changes
for catalog.schema.table names, we can stop doing the monkeypatch,
so get rid of all that, and re-enable testing during a build.

The tests used to fail because of the catalog.schema.table change.

So overall, pay off some debt!