[DM-17187] Add DISTANCE to capabilities.
Now that we have DISTANCE working, add it to the capabilities.  This
is just nice, because some clients (like TOPCAT), check and compare
the capabilities against the query being executed.

Without this change, TOPCAT warns the user that DISTANCE isn't supported
by the endpoint (although it lets you send it anyway).  This just gets
rid of that error.