[DM-17030] Get Travis passing again.
First, add the user name and password for oracle maven's
repository into the travis job (not in the .yml though).

With those set, now it's just editing the docker-compose
script, which is run in travis to set up an environment.
That required a few changes, like pointing tap_schema to
a different database container, to make it more like how
we run it in k8s.

Also this required a small change to query a table other
than the obscore table, which is in flux right now.