[DM-20729] Parse Bearer Token
In the AuthenticatorImpl class, which we tie in using a system
property passed in at runtime to point to our class.  This class
will look for a BearerTokenPrincipal and then decode it into
other principals with the user id and name to use in other
parts of the code.

We decode the token using a 3rd party library that handles these
sorts of things.