[DM-20729] Auth fixes
First, on the opencadc/core PR, there have been some changes, like
getting rid of getToken.  So AuthenticatorImpl was changed to use
the new API.  Those changes also require a new version, which is
marked in the build.gradle.  That's the version of the new PR.

The QueryJobManager was using the ACIdentityManager, which wasn't
exactly correct and caused a problem when persisting it the info
to the database for owners.  But thanks to @bmajor, he pointed out
that we should be using the X500IdentityManager, which doesn't
have this problem.