[DM-21826] Fix travis breakage: owner_id column
So when CADC added the table and schema permissions for uploading
user tables and data, they added fields to the tap_schema database
tables to track the owner.  These are required, and if they don't
exist, then the query will fail when looking at the tap_schema,
since it looks at these columns by name.

This was somewhere around this commit:
https://github.com/opencadc/tap/commit/790833901ab7ab40dec5162e1225426c3f3ab7eb

But we never added these columns to the tap_schema database, so
for new versions that use this version of the TAP library, we need
to update the database for these.  So I'm editing the tap_schema
creation script to add those in.

Also, we need to specify the identitymanager class in the docker
compose part (or at least, that is how it runs in production, so it
is good to test it that way).