[DM-22013] Cleanup build step
First off, get more consistent about the names of Dockerfiles
and how they relate to the container name they are building.

Second, make the build scripts reflect those, using the better
names.  Allow for building a dev tag locally, then pushing it
to whatever name provided on the command line.

Third, get rid of unused files.  Most of these were unused or
only used for presto, which we aren't currently using.  If we
decide to use these in the future, they are always in the history
for resurrecting.