[DM-31776] Initial example for working datalink
So first you have to edit the datalink-manifest.json to say
what the name of the xml file is that contains the snippet,
as well as what columns are required to include that snippet.

The snippet is then passed through the StringTemplate library
to put in the proper column number references based on the columns
that are returned to the user.