[DM-31965] Datalink: allow templating of url
This will help us pass in the hostname that is different between
each environment, so it should help us reduce the amount of configs.

This is already available in the chart, since it's needed by the
ingress.

This isn't the way it is done in the capability endpoint, which
apparently looks at the request for the hostname that was sent
in on, and using that in the response.  Since we don't have access
to the request at this point, we can't do that.