Redo error handling in new S3 support code
Turn errors in constructing a URI from the bucket URL into an
unchecked exception to avoid a try/catch block that prints a
backtrace and otherwise ignores the error.  This will hopefully
crash TAP, which is reasonable given that this is a configuration
error, rather than trying to continue and producing hard-to-diagnose
errors.

Also fix spacing and indentation style in a few places.