Overhaul image building with GitHub Actions
Use the standard GitHub Actions to build and push the Docker
images rather than hand-rolled scripts.  Update the tag script
to the latest version from the Safir FastAPI template.  Also push
images to the GitHub Container Registry, since we're slowly moving
off of Docker Hub.

This requires changing the build to be relative to the top-level
directory instead of the docker directory, since the GitHub Actions
don't appear capable of dealing with the latter.