[DM-37254] Emit exception message as warning
Right now this is printed out at the debug level, but it will be a
lot easier to find and get a hold of if we make it a warning.  Since
this is an exception path, this shouldn't really add too much log
spam, considering the rethrown messages are emitted after, but we
seem to lose the first one which is the most important.